Fixup libusb_exit to be more thread-safe
Be more careful about locks in libusb_exit

Acquire the global lock, check if global state is being referenced.
Only if it's not do we release it (i.e. don't read global state without
lock acquired).

Move the de-initialization of the global context to the end because the
de-init parts of code may reference the default context.