Fix potential deadlock in the thread shutdown scenario.
I've seen CFRunLoopStop not always stop the event thread.
According to the docs that is possible as the run loops may
be nested & the stop will only stop the outermost one.

This isn't ideal, but should ensure this issue doesn't actually
pop-up (& if it does, handle it more gracefully than deadlocking).