Implement libusb_get_portpath on OSX
Make use of the existing sys_path member variable of the private
data structure.  It's not being used for anything anyways.

Additionally, change the construction of the string to be an actual
port topology path instead of being an instance ID.  This is useful to
keep track of a device on a port that may go through several modes with
different VID/PID & descriptors.