Add extended pollfd notification API
On Windows, the "pollfd" to report up to the user is actually an event
HANDLE, so we add an extended API that reports to the callback the
appropriate value.

According to my understanding of MSDN, HANDLE's are
actually 32-bit safe values for compatability reasons, so there
shouldn't be a problem with thunking to an integer (although eventually
using ssize_t is better if we can break the API & ABI).